pyproject.toml
[tool.poetry]
name = "nasy_night_bot"
version = "0.1.0"
description = "Nasy Night Telegram Bot"
authors = ["Nasy <nasyxx+git@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
